Fix for ticket 156.
Modify ParseFunction() to check for Spell() and SpellCooldown() statements
in the script and check if the given spell ID is in the spellbook or not.
If it's not, then see if it matches another spell with the same name that
is in the spellbook.  If so, then add that spell ID to the OvaleData
spellbook.

Track the function nodes with unknown spell IDs in a list and tag the
unknown spell IDs with ":unknown" when invoking DebugNode().

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@520 d5049fe3-3747-40f7-a4b5-f36d6801af5f