Fix for ticket 176.
The Ovale icon bar is being shown/hidden improperly when the player is in
a pseudo-vehicle where the player is controlling something other than
himself.  Use UnitHasVehicleUI() instead of UnitInVehicle() to detect this
condition properly.  UnitHasVehicleUI() returns true if the player has a
vehicle action bar, which corresponds to the player not being able to use
his own abilities.

Analysis and fix from Galvin in ticket 176.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@599 d5049fe3-3747-40f7-a4b5-f36d6801af5f