Split compilation into two phases: "quick" and "full".
The "quick" phase only compiles the parts of the script that may change as
a result of things that trigger recompilation.  This includes changing
SpellInfo(), AddCheckBox() and AddListItem().

The "full" phase compiles the entire script.  This has been modified to
also compile AddIcon() groups that may be disabled and storing their nodes
in Ovale.masterNodes.

At run-time, when a "quick" compile is requested, Ovale executes the quick
compile and then updates the status of each master node (corresponding to
an icon) as to whether the node is enabled or not.

OvaleFrame only shows icons for enabled master nodes.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@604 d5049fe3-3747-40f7-a4b5-f36d6801af5f