Remove split compilation and use a single pass for compiling again.
Back out part of r604 that split compilation into a "quick" part and
an expensive part.  The more efficient method in r606 for searching
and replacing defines with key values seems to have fixed the slow
compilation enough that it fits within the Lua execution window.

Re-merging the compilation phases fixes some issues with Ovale not
properly updating the script when a toon levels, as well as some
other edge cases.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@636 d5049fe3-3747-40f7-a4b5-f36d6801af5f