Fix for ticket 198 - Wrong SpellID for bloodbath.
The spell ID generated by the SimC converter is incorrect.  113344 is the
spell ID of the debuff that is applied to the targets hit by the warrior
when the warrior has the buff with ID 12292 that is triggered by the spell
with the same ID.

Also fix the SpellInfo and SpellAddBuff lines associated with Bloodbath to
match the buff and not the debuff.

This matches the usage of the "bloodbath" references in the SimC profile,
which refer to "buff.bloodbath.up".

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@669 d5049fe3-3747-40f7-a4b5-f36d6801af5f