Add Tier 14 gear for all classes.
Only tier gear whose set bonuses affect either the cooldown or duration of
spells are tracked.

Make the names of the Tier 14 item sets for hybrid classes more like the
names used in SimC by appending one of the following suffices to the
standard "T14" name:

	_caster, _heal, _melee, _tank

This allows for differentiating, for example, between a druid wearing 4
pieces of feral gear versus 4 pieces of guardian gear.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@678 d5049fe3-3747-40f7-a4b5-f36d6801af5f