Don't fill the spellbook with missing spells before compiling new script.
Before this change, the spellbook (OvaleData.spellList) would be populated
with entries found from a previous script compilation, which may have
changed after the event SPELLS_CHANGED is received.

This change makes filling the spellbook with missing spells found in the
script happen *after* the script is compiled, and moves the responsibility
into OvaleCompile:Compile().

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@681 d5049fe3-3747-40f7-a4b5-f36d6801af5f