Cache active glyphs in OvaleData and update on GLYPH_* events.
This change causes the OvaleData module to update the active glyph cache
on GLYPH_* events.

All modules now reference the OvaleData.glyphs table instead of looping
through all of the glyph sockets and querying their contents each time we
need to find out if a particular glyph is active.

This is an efficiency boost for scripts.

Also fix "/ovale glyph" to output the currently active glyphs.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@704 d5049fe3-3747-40f7-a4b5-f36d6801af5f