Fix ticket #211 - Burning Ember Number
UnitPower() has a hidden 3rd argument to get the true value of a warlock's
secondary power value.  This fixes BurningEmbers() and SoulShards() to
return values from 0 to 40 (for the number of ember or shard segments)
instead of 0 to 4 (full ember or shard).  This also now correctly
correlates with the burningembers=xxx and shards=xxx values in warlock
script SpellInfo() lines.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@710 d5049fe3-3747-40f7-a4b5-f36d6801af5f