Add new SpellInfo parameters to handle buffs generating extra resources.
``buff_<resource>'' is an optional parameter that is set to the spell ID
of the buff that causes extra resources to be generated when present.

``buff_<resource>_amount'' is an optional parameter that is set to the
number of extra resources generated when the buff listed in
``buff_<resource>'' is present.  If not set, then it defaults to ``1''.

Drop special handling of Shadow Blades in OvaleState now that it can be
described and handled directly in the rogue script with the above
parameters.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@712 d5049fe3-3747-40f7-a4b5-f36d6801af5f