Better encapsulation of compile-related code into OvaleCompile module.
The script needs to be recompiled if data that the script depends upon
changes.  In the other modules that flagged ``needCompile'', cause the
data changes due to Blizzard events to fire a new Ovale event after the
data is changed.

The new Ovale events are:

    Ovale_CheckBoxValueChanged
    Ovale_EquipmentChanged
    Ovale_GlyphsChanged
    Ovale_ListValueChanged
    Ovale_ScriptChanged
    Ovale_SpellsChanged
    Ovale_StanceChanged
    Ovale_TalentsChanged

Make OvaleCompile into a full-fledged Ace3 addon module that listens for
the new Ovale events that triggers script recompilation.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@714 d5049fe3-3747-40f7-a4b5-f36d6801af5f