Add new module OvaleScripts to manage scripts from multiple sources.
Move the existing default scripts into OvaleScripts and decorate them with
a description that allows a user to see what class specializations are
supported by the default script.

Teach Ovale to use the script ``source'' to find the code for the current
script.  Custom code can be copied from an existing script to a blank
custom template that may be modified and saved into the SavedVariables
file.  The script ``source'' may be selected from a dropdown menu, with
the default menu containing the Ovale default script and either a blank
custom script or the custom script loaded from the SavedVariables file.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@715 d5049fe3-3747-40f7-a4b5-f36d6801af5f