Improve encapsulation of OvaleFuture.
Add new method to OvaleFuture that generates an iterator for other modules
to walk the list of spells that are cast or are in flight to their
targets.  Make the list of active spells private since it doesn't need to
be accessed outside of OvaleFuture.

Use the iterator in OvaleState to apply the spell effects of those
travelling spells, and remove OvaleFuture:Apply() since it is no longer
needed.  This de-couples OvaleFuture from OvaleState.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@756 d5049fe3-3747-40f7-a4b5-f36d6801af5f