Warrior: Apply fixes after SimulationCraft import.
The spell ID generated by the SimC converter is incorrect.  46915 is the
spell ID of the passive ability while 46916 is the spell ID of the actual
buff gained by the warrior.

Also add a SpellInfo line that notes the duration of the Bloodsurge buff.

This matches the usage of "bloodsurge" in the Ovale script which checks
for BuffStacks(bloodsurge), and which also has a SpellAddBuff line for
wild_strike that deducts one from Bloodsurge stack.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@767 d5049fe3-3747-40f7-a4b5-f36d6801af5f