Mage: Apply fixes after SimulationCraft import.
Add damage coefficients to Fireball, Inferno Blast and Pyroblast so that
the Damage() and CritDamage() conditions used in deciding when to use
Combution return sensible numbers.

Fix confusion between ``pyroblast'' and ``pyroblast_aura''.  The former is
the ID of the spell as well as the DoT left on the target hit by
Pyroblast.  The latter is the player buff that indicates the next
Pyroblast is instant-cast.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@768 d5049fe3-3747-40f7-a4b5-f36d6801af5f