Code consistency.
Make local copies of all Blizzard API functions and prefix them with
"API_". This makes it simpler to automatically audit the use of Blizzard
API functions.

Rename all private static variables by adding a "self_" prefix.  This
reduces namespace conflicts with variable names when writing functions.

Rename all constants and constant tables to be in all-capitals and to
start with an "OVALE_" prefix.  This makes it clear what values are
immutable and which may be changed.

Also remove ``Ovale'' from the global namespace as a final clean-up.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@797 d5049fe3-3747-40f7-a4b5-f36d6801af5f