Attempt to fix ticket #191 - "Other" condition not working as expected.
Change GetExpirationTimeOnAnyTarget() methods in OvaleAura and OvaleState
into a new method GetAuraOnAnyTarget().  It returns a time interval
similar to the GetAura() methods and takes similar arguments as well.

Use the new method in OvaleCondition and change semantics of the private
method getOtherAura() to return the same results as GetAuraOnAnyTarget().

Change Other*() conditions to use the return value from getOtherAura()
properly.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@805 d5049fe3-3747-40f7-a4b5-f36d6801af5f