Add GetDamageMultiplier(spellId) method to OvaleState.
This method accounts for auras that are removed by spells that are still
in flight to their targets.  Use this method in place of the one from
OvaleAura in OvaleCondition since OvaleCondition should only reference the
current frame state.

Change the implementation in OvaleAura to not maintain the current damage
multiplier while scanning the auras and to try to account for stacking
auras.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@814 d5049fe3-3747-40f7-a4b5-f36d6801af5f