Convert OvaleBestAction:Compute() to use a lookup table.
This change converts the Compute() method of OvaleBestAction to use a
lookup table instead of cascading if-else statements for efficiency.  As a
side-effect, it is also easier to understand and maintain the code this
way.

Re-add node IDs to each node of the abstract syntax tree in OvaleCompile.
Node IDs are used in the debugging messages for OvaleBestAction.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@816 d5049fe3-3747-40f7-a4b5-f36d6801af5f