Add rate-of-change to ``Health'', ``HealthMissing'', and ``HealthPercent''.
The rates are calculated from the current health and maximum health of the
specified unit and the estimated DPS against that unit.  This makes the
``Health'', ``HealthMissing'', and ``HealthPercent'' conditions return a
time interval that can be used by OvaleBestAction to compute sensible time
intervals, e.g., ``target.HealthPercent() <25'' has a time estimate
associated with it.

Change the private TimeToDie() method to return a triplet of:

    time to die, current health, maximum health

to support the new calculations.

Enhance the private TimeToDie() method to be more accurate within the
first few seconds of targeting a new mob.

Change the ``DeadIn'' condition to be an alias of the ``TimeToDie''
condition.

Change Life/LifeMissing/LifePercent to be aliases of
Health/HealthMissing/HealthPercent instead of vice versa, as the latter
are names that correspond to the SimulationCraft Action List grammar and
are more likely to be looked up.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@818 d5049fe3-3747-40f7-a4b5-f36d6801af5f