Enhance GetAura methods to accept the name of a spell list.
This allows for doing things like:

    OvaleAura:GetAura("player", "melee_haste", "HELPFUL")
    OvaleState:GetAura("player", "melee_haste", "HELPFUL")

to get the first buff from the melee_haste spell list that is found on the
player.

As a side-effect of this change, the SpellInfo parameters that take a
buff spell ID, e.g., buffnocd, buff_combo, etc., can now take a spell list
name instead.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@822 d5049fe3-3747-40f7-a4b5-f36d6801af5f