Save the tick length as part of the aura properties.
Set the tick information directly in AddAura method instead of the CLEU
event handler in case CLEU and UNIT_AURA events arrive out of order.

Adjust the API by changing methods that returned the spell haste
multiplier to return the tick length instead.  The primary use of the
spell haste multiplier was to compute the tick length later on anyway.

Drop the hasteMultiplier argument to GetDuration and GetTickLength since
those two methods are now never called with that argument set.

Enhance GetTickLength and GetDuration to accept spell list name in place
of spell ID.

Note that the various ``tick'' conditions can take a spell list.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@834 d5049fe3-3747-40f7-a4b5-f36d6801af5f