Defer creating strings until the actual Print/Log/Error methods.
Add new methods to Ovale -- Logf, Errorf, DebugPrintf -- that take a
format string and arguments for interpolation, much like the standard
AceConsole-3.0 Printf method.  Use the new methods in place of Log, Error,
DebugPrintf so that the actual creation of strings is deferred until the
arguments are interpolated into the format string within the new methods.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@843 d5049fe3-3747-40f7-a4b5-f36d6801af5f