Back out changes in r842 and restore damage meter registration.
OvaleFuture needs to send out messages to all of the registered
damage meters when a scored spell has been cast.

Improve the implementation of the registration so that it only
requires a function or a method name instead of the module itself.
This makes it easier for other damage meters to implement a small
function to receive score information from Ovale instead of
implementing a full-blown Ovale module.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@855 d5049fe3-3747-40f7-a4b5-f36d6801af5f