Track critical strike chance and split bonus damage from bonus healing.
This isn't perfect.  This only updates the player's critical strike chance
if it's due to a change in critical strike rating.  If there is a buff
applied that provides a flat increase to a player's critical strike
chance, then it won't trigger the update.

We also use two different events for updating bonus damage and bonus
healing instead of just one, which may not necessarily have fired when
bonus healing changed.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@876 d5049fe3-3747-40f7-a4b5-f36d6801af5f