Dynamically update tick length of auras applied by the player.
The tick times of a periodic aura applied by the player are taken from
watching for SPELL_PERIODIC CLEU events.  Those tick times are used to
update the estimate for the aura's tick length based on empirical
evidence, and is hopefully more accurate.  It also avoids needing to store
the haste percentage at which a periodic aura was applied.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@897 d5049fe3-3747-40f7-a4b5-f36d6801af5f