Don't account for stacking damage buffs for now.
Back out part of r814 that multiplied the damage increase by the stack
size.  There are very few stacking damage buffs, and those can be handled
explicitly for now until Ovale gains a way to describe a stacking versus a
non-stacking buff.

This fixes the result of the Damage() and LastSpellEstimatedDamage()
conditions.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@912 d5049fe3-3747-40f7-a4b5-f36d6801af5f