Change internal implementation of spell lists.
Move internal representation of spell lists from lists whose values are
the spell IDs to tables indexed by spell IDs with a default value of
"true".  This makes it easier to do the most common operation, which is to
check whether a given spell ID is in the spell list -- we can now do a
simple table lookup instead of walking the list.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@930 d5049fe3-3747-40f7-a4b5-f36d6801af5f