Extend tracking of auras cast by others to track all auras.
The assumption is that it's not possible for one caster to put more than
one aura of a given ID on a target, which should be true.

Change GetMyAuraOnAnyTarget to GetAuraOnAnyTarget, which allows for
specifying whether to check for only "my" auras or auras cast by anyone.
With the associated change to extend aura tracking above, the following
script conditions should provide more accurate answers:

    BuffCount, DebuffCount
    OtherBuffExpires, OtherDebuffExpires
    OtherBuffPresent, OtherDebuffPresent
    OtherBuffRemains, OtherDebuffRemains

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@938 d5049fe3-3747-40f7-a4b5-f36d6801af5f