Don't cheat and stash an extra "Magic/Disease/etc." aura.
Remove the workaround that stashed an extra aura whose spellId was the
debuff type of the aura.  This was to allow for implicit spell lists named
"Magic", "Poison", "Disease", and "Curse" that could be used where other
spell list names are used.

We now just check all of the auras on the target to see if there is one of
the named debuff type on the target.

As part of the cleanup, rename some "spellId" variable names to "auraID"
to reduce namespace collision.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@939 d5049fe3-3747-40f7-a4b5-f36d6801af5f