Snapshot player stats when a spell is cast and when DoTs are applied.
The most recent successfully landed cast of a spell is saved in a table
indexed by target and spell ID, and contains all of the information for
that spellcast, including cast time and also a snapshot of the player
stats.

Player stats are snapshotted both at the start of a cast and when it is
sent to the server so that OvaleState can approximately know the player's
stats at the end an active spellcast.

Player stats are also snapshotted when a DoT is applied or refreshed.

Introduce a new generalized public method OvaleFuture:GetLastSpellInfo()
that returns information about the last successfully landed spell.  Remove
the following superseded methods:

	GetLastAttackPower
	GetLastComboPoints
	GetLastCritChance
	GetLastDamageMultiplier
	GetLastSpellpower

Change the following conditions to take a target since spellcast
information is also tracked by target:

	LastSpellAttackPower
	LastSpellComboPoints
	LastSpellCritChance
	LastSpellDamageMultiplier
	LastSpellEstimatedDamage
	LastSpellMastery
	LastSpellSpellpower

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@941 d5049fe3-3747-40f7-a4b5-f36d6801af5f