Add new script conditions.
The following two conditions allow for inspection of the player's current
stats:

    MeleeCritChance
    RangedCritChance

The following script conditions allow for examining the player's stats at
the time the given DoT was applied on the target.

    DebuffAttackPower
    DebuffSpellpower
    DebuffMeleeCritChance
    DebuffRangedCritChance
    DebuffSpellCritChance
    DebuffSpellHaste
    DebuffMastery

This information can be used to determine whether or not to overwrite an
existing DoT on the target.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@943 d5049fe3-3747-40f7-a4b5-f36d6801af5f