Fix ticket 256 - DebuffAttackPower not updating properly on refresh
Determining whether an aura needs a fresh snapshot requires some state to
be kept.  In this case, we track the most recent spell cast by the player
that preceded the aura being applied.  If that spell is one that doesn't
trigger a new snapshot of the aura, then don't snapshot; but snapshot
otherwise.

Currently doesn't deal with DoTs that are spread by other spells, e.g.,
Inferno Blast and Soul Swap.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@969 d5049fe3-3747-40f7-a4b5-f36d6801af5f