Separate the base damage multiplier from the spell-specific one.
The "base damage multiplier" is the one on the character sheet and applies
to all attacks -- it is referenced as "baseDamageMultiplier" in the
player's stat snapshot.

The "spell-specific damage multiplier" for a spell is an extra damage
multiplier for that spell, excluding the base effects, that includes
effects from buffs, debuffs, and stances -- it is referenced as
"damageMultiplier" in the spellcast and aura information.

Move OvaleAura:GetDamageMultiplier into OvaleFuture and make it private
since OvaleFuture is the only one that uses it.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1007 d5049fe3-3747-40f7-a4b5-f36d6801af5f