Fix ticket 296 - Ovale double deducting for energy cost
Only apply the power cost of the spell if the spellcast hasn't already
finished.  A spellcast that has finished in-game has already had the power
adjusted, which would be picked up when :Reset() is called to reset the
frame state.

Also, make this more readable by splitting out code from :ApplySpell()
method to update the power state into a new method :ApplySpellCost().

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1036 d5049fe3-3747-40f7-a4b5-f36d6801af5f