Workaround for ticket 290 - Dream of Cenarius Second Charge Rip.
When a buff expires simultaneous with a spellcast, a debuff applied by
that spellcast may not snapshot the effects of the expired buff.

Tag an expired buff as "consumed" if the aura ended early and the ending
cast time of the previous spellcast is within 0.3s of the expiration time
of the aura.  Treat an aura as still "active" if it was consumed and we
are checking within 0.3s of when it has expired.  This allows an aura that
provides a damage multiplier to still be added into the snapshot for any
applied auras.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1238 d5049fe3-3747-40f7-a4b5-f36d6801af5f