Most buff script conditions should be allowed to return non-active auras.
The script evaluation engine (OvaleBestAction) handles expired auras just
fine since it works in intersections and unions of time spans, but this
will allow the engine to properly deal with auras that are applied in the
future, e.g., after the current spellcast.  This fixes the balance druid
scripts that recommend applying DoTs once an Eclipse state is reached.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1252 d5049fe3-3747-40f7-a4b5-f36d6801af5f