Fix bug in r1206 in the way that slot <--> action mappings are managed.
The spell/item/macro to slot ID mapping wasn't correctly being added due
to an improper condition.  Now add the mapping if it either doesn't
already exist or if the new slot ID is lower than the existing one (in the
case where a spell is present multiple times on the action bar).

This fixes the return values of the :GetForSpell(), :GetForMacro(), and
:GetForIem() methods to return something other than nil.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1298 d5049fe3-3747-40f7-a4b5-f36d6801af5f