Implement RuneOfPowerRemains() condition.
This translates SimC's "buff.rune_of_power.remains" by returning the
remaining time of the most recently-placed Rune of Power if the player is
currently standing within a Rune of Power.

The assumption here is that the player is standing in the most
recently-placed Rune of Power.

This fixes ticket #341.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1320 d5049fe3-3747-40f7-a4b5-f36d6801af5f