Allow for expressing "if glyph NOT active" or for other parameters.
SpellInfo parameter glyph=id currently means that the declaration is true
only if the given glyph is active.  This change allows for the opposite
meahing by using glyph=!id to mean that the declaration is true only if
the given glyph is NOT active.

Other parameters that allow for the key=!value syntax include:

    glyph, mastery, if_stance, if_spell, talent

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1365 d5049fe3-3747-40f7-a4b5-f36d6801af5f