Fix and extend the tooltips shown when hovering over an icon.
SVN revision r1117 -- that implemented a true timespan implementation --
broke the tooltip that showed the name of the spell being shown when
hovering over an icon.  Fix the tooltip to show the name of the spell
again and extend the facility to allow for showing item or macro names in
the future.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1370 d5049fe3-3747-40f7-a4b5-f36d6801af5f