Treat all auras as existing from gain time to ending time.
Due to DoT refresh mechanics, the gain time of an aura may be before the
start time of the aura (when it starts ticking again).  Treat all auras as
existing from the gain time to expiration time.  A refreshed aura still
exists when it is gained, even though it starts ticking later.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1383 d5049fe3-3747-40f7-a4b5-f36d6801af5f