Make TickTime() return the tick length of a new DoT if one isn't present.
This matches SimC semantics where if a DoT is present, then "tick_time" is
the actual tick length of the DoT, but if the DoT is not present, then it
is the tick length of a new DoT with a current stat snapshot.

This fixes mage scripts not suggesting mage bombs.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1390 d5049fe3-3747-40f7-a4b5-f36d6801af5f