Workaround for aura lag.
This is ultimately a problem in OvaleFuture where a spell may have
finished casting and landed on the target, but any target auras applied by
the spell haven't yet been applied on the target due to aura lag.

Need to find a way to still apply spell effects on a spell that has landed
on its target but not yet applied auras.  Perhaps split out another level
of ApplySpell() due to aura lag.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1411 d5049fe3-3747-40f7-a4b5-f36d6801af5f