Teach OvaleCooldown about spell charges.
Lazily update the cooldown information of spells based on the age of the
cooldown state, which is flagged to be updated when a spell's charges or
cooldowns is changed or if a spell is successfully cast.

Add a GetSpellCharges() method to the state machine to act like the
Blizzard API function of the same name.

Fixes ticket 357.

Age cooldown state properly.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1419 d5049fe3-3747-40f7-a4b5-f36d6801af5f