Add additional phase of tracking to spellcasts: "after hit".
ApplySpellAfterHit() is called after the three previous phases:

    ApplySpellStartCast
    ApplySpellAfterCast
    ApplySpellOnHit
    ApplySpellAfterHit

Apply spell auras to the target during ApplySpellAfterHit().  Only remove
a spellcast from the spells-in-flight after it has applied any auras to
the target, which can take a bit of time due to server event lag.  The
simulator will continue to apply spells until it sees their aura effects
appear on the target in reality.

Remove the previous workaround for aura lag added to OvaleAura in r1411
and r1413.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1426 d5049fe3-3747-40f7-a4b5-f36d6801af5f