Back out r1484 that added Outbreak to 1H Frost rotation.
Outbreak and Unholy Blight are not the same priority.  Even though they
both share the same GCD cost, the latter will apply diseases to everything
around the DK while Outbreak only applies them to the current target.
Outbreak was removed from the 1H Frost rotation in SimulationCraft as part
of an optimization for MoP 5.4.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1486 d5049fe3-3747-40f7-a4b5-f36d6801af5f