Enhance some methods for checking weapons to allow for Titan's Grip.
Titan's Grip allows for a 2H weapon to be equipped in the off-hand.  Add
extra parameters to HasOneHandedWeapon() and HasTwoHandedWeapon() to allow
checking in a specific weapon slot, and to HasMainHandWeapon() and
HasOffHandWeapon() to allow checking explicitly for either a 1H or 2H
weapon.

Give HasWeapon() condition an extra paramter "type=value" where value is
"1h" or "2h" to allow checking explicitly for a 1H or 2H weapon in the
given hand.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1490 d5049fe3-3747-40f7-a4b5-f36d6801af5f