Add new Profiler module to allow basic profiling of Ovale code.
To get profiling stats for a function, execute Profiler:Start(tag) at the
head of the function and Profiler:Stop(tag) before exiting the function,
where "tag" is a unique label to track.

Everything that happens in-between Start/Stop is attributed to the given
tag.  If a profiled function calls another profiled function, then the
time spent within the called function is not charged to the calling
function.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1504 d5049fe3-3747-40f7-a4b5-f36d6801af5f